find bug with itf enum; C starts STA at zero, we started at 1 :(
